Updated time_interval
Updated: time_interval to be called universal_time_interval, and the
interval_measure of it to be called universal_time, not to be confused
with the time of a stop watch.  The id of the subdim "date", was changed
to "day", as some data formats use the Date format as equivalent with
DateTime, which includes hours and minutes etc. And the pattern for week
and quarter, needs to have a w & q (lower case) in them, to ensure they
are unique