add a helper script to filter diff result
I added the diff.py so that we can filter the result return by daff.

usage:

- python diff.py path/to/diff.txt path/to/output.txt

you can change the threshold in diff.py to get different results.

the diff-filtered.txt is filtered with a 10% threshold, i.e if value in SG more
than 10% different to GW, the program keep the line

if value in GW is 0, then if SG is bigger than threshold / 100, the program will
keep the line; and if SG and GW have different sign, it will keep the line too.